+++ src/core/exchanges/binance/exchange_client.rs
@@ -114,22 +114,6 @@ impl ExchangeClient for Binance {
         Ok(())
     }
 
-    async fn request_open_orders_by_http_header(
-        &self,
-        http_params: Vec<(String, String)>,
-    ) -> Result<RestRequestOutcome> {
-        let url_path = match self.settings.is_margin_trading {
-            true => "/fapi/v1/openOrders",
-            false => "/api/v3/openOrders",
-        };
-
-        let full_url = rest_client::build_uri(&self.settings.rest_host, url_path, &http_params)?;
-
-        let orders = self.rest_client.get(full_url, &self.settings.api_key).await;
-
-        orders
-    }
-
     async fn request_open_orders(&self) -> Result<RestRequestOutcome> {
         let mut http_params = rest_client::HttpParams::new();
         self.add_authentification_headers(&mut http_params)?;
